Move isValidDate and associated test from time to date files
